added acc impl

+++ src/main/java/com/springbootdhan/dao/AccountDao.java
@@ -4,4 +4,4 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.springbootdhan.entity.Account;
 
-public interface AccountDao extends JpaRepository<Account, Integer>{}
+public interface AccountDao extends JpaRepository<Account, Long>{}
